requirements.txt
openai
python-dotenv
langchain
langchain-openai
langchain-community
chromadb
langchain-chroma
pypdf
tiktoken 
unstructured
pytesseract
Pillow
python-pptx
python-docx 
nltk 